# api/forgeworks_api.py
"""
REFORGE OS - ForgeWorks Core API
FastAPI server that integrates Rust services for compliance-first device analysis
"""

from fastapi import FastAPI, HTTPException, Body, Query, Header
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse, FileResponse
from typing import Optional, Dict, Any, List
from pydantic import BaseModel
import json
import os

# Import Rust services (via FFI or HTTP - for now, we'll use Python wrappers)
# In production, these would call Rust via FFI or HTTP

app = FastAPI(
    title="REFORGE OS - ForgeWorks Core",
    description="Compliance-first device analysis and routing platform",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# ============================================================================
# REQUEST/RESPONSE MODELS
# ============================================================================

class DeviceAnalyzeRequest(BaseModel):
    device_metadata: str
    platform: Optional[str] = None
    connection_state: Optional[str] = None

class DeviceAnalyzeResponse(BaseModel):
    ok: bool
    device_id: str
    model: str
    manufacturer: str
    security_state: str
    capability_class: str
    classification: str
    restrictions: List[str]
    non_invasive: bool = True

class OwnershipVerifyRequest(BaseModel):
    user_id: str
    device_id: str
    attestation_type: str  # PurchaseReceipt, CourtOrder, etc.
    documentation_references: List[str] = []

class OwnershipVerifyResponse(BaseModel):
    ok: bool
    verified: bool
    confidence: float  # 0.0 - 1.0
    required_authorization: Optional[str] = None
    blocked: bool

class LegalClassifyRequest(BaseModel):
    device_id: str
    ownership_confidence: float
    jurisdiction: str  # US, EU, UK, etc.

class LegalClassifyResponse(BaseModel):
    ok: bool
    status: str  # Permitted, ConditionallyPermitted, Prohibited, RequiresAuthorization
    jurisdiction: str
    authorization_required: List[str]
    risk_level: str  # Low, Medium, High, VeryHigh
    routing_instructions: Dict[str, Any]

class ComplianceSummaryRequest(BaseModel):
    device_id: str
    include_audit: bool = True

class ComplianceSummaryResponse(BaseModel):
    ok: bool
    device: Dict[str, Any]
    ownership: Dict[str, Any]
    legal: Dict[str, Any]
    routing: Dict[str, Any]
    audit_entries: List[Dict[str, Any]]
    report_timestamp: str
    audit_integrity_verified: bool

class InterpretiveReviewRequest(BaseModel):
    device_id: str
    scenario: str
    ownership_confidence: float

class InterpretiveReviewResponse(BaseModel):
    ok: bool
    classification: str
    risk_framing: Dict[str, Any]
    historical_context: str  # High-level only, no instructions
    authority_paths: List[Dict[str, Any]]
    compliance_notes: str

# ============================================================================
# MOCK IMPLEMENTATIONS (Replace with Rust FFI/HTTP calls)
# ============================================================================

def mock_device_analyze(metadata: str) -> DeviceAnalyzeResponse:
    """Mock device analysis - replace with Rust service call"""
    # In production: Call Rust device_analysis::analyze()
    return DeviceAnalyzeResponse(
        ok=True,
        device_id="dev_001",
        model="iPhone 13 Pro" if "iPhone" in metadata else "Galaxy S21",
        manufacturer="Apple" if "iPhone" in metadata else "Samsung",
        security_state="Analyzed - Clean",
        capability_class="No modifications detected",
        classification="Clean",
        restrictions=["No modification capability", "Analysis only", "Read-only operations"],
        non_invasive=True
    )

def mock_ownership_verify(req: OwnershipVerifyRequest) -> OwnershipVerifyResponse:
    """Mock ownership verification - replace with Rust service call"""
    # In production: Call Rust ownership_verification::verify_ownership()
    confidence_map = {
        "CourtOrder": 0.95,
        "ServiceCenterAuthorization": 0.90,
        "EnterpriseAuthorization": 0.85,
        "PurchaseReceipt": 0.80,
        "InheritanceDocument": 0.75,
        "GiftDocument": 0.70,
    }
    confidence = confidence_map.get(req.attestation_type, 0.50)
    verified = confidence >= 0.85
    blocked = confidence < 0.50
    
    return OwnershipVerifyResponse(
        ok=True,
        verified=verified,
        confidence=confidence,
        required_authorization="OwnershipProof" if not verified else None,
        blocked=blocked
    )

def mock_legal_classify(req: LegalClassifyRequest) -> LegalClassifyResponse:
    """Mock legal classification - replace with Rust service call"""
    # In production: Call Rust legal_classification::classify_legal_status()
    if req.ownership_confidence < 0.50:
        status = "Prohibited"
        auth_required = ["CourtOrder", "ServiceCenterAuthorization"]
        risk = "VeryHigh"
    elif req.ownership_confidence < 0.85:
        status = "RequiresAuthorization"
        auth_required = ["OwnershipProof"]
        risk = "Medium"
    else:
        status = "Permitted"
        auth_required = []
        risk = "Low"
    
    return LegalClassifyResponse(
        ok=True,
        status=status,
        jurisdiction=req.jurisdiction,
        authorization_required=auth_required,
        risk_level=risk,
        routing_instructions={
            "route_to": "OEM" if status == "Permitted" else "LegalCounsel",
            "contact_information": "Contact legal counsel for jurisdiction-specific guidance",
            "required_documentation": ["Ownership proof", "Authorization documents"],
            "compliance_notes": "External authority routing required" if status != "Permitted" else "Standard routing"
        }
    )

def mock_compliance_summary(device_id: str) -> ComplianceSummaryResponse:
    """Mock compliance summary - replace with Rust forgeworks_core::process_device_flow()"""
    # In production: Call Rust forgeworks_core::process_device_flow()
    return ComplianceSummaryResponse(
        ok=True,
        device={
            "device_id": device_id,
            "model": "iPhone 13 Pro",
            "security_state": "Analyzed - Clean",
            "non_invasive": True
        },
        ownership={
            "verified": True,
            "confidence": 0.90,
            "blocked": False
        },
        legal={
            "status": "Permitted",
            "jurisdiction": "US",
            "risk_level": "Low"
        },
        routing={
            "route_to": "OEM",
            "compliance_notes": "Standard routing"
        },
        audit_entries=[
            {
                "action": "device_analysis",
                "result": "Allowed",
                "timestamp": "2025-01-10T00:00:00Z"
            }
        ],
        report_timestamp="2025-01-10T00:00:00Z",
        audit_integrity_verified=True
    )

def mock_interpretive_review(req: InterpretiveReviewRequest) -> InterpretiveReviewResponse:
    """Mock interpretive review - replace with Rust + Pandora Codex (internal)"""
    # In production: Call Rust services + Pandora Codex risk models (internal only)
    # Pandora Codex informs risk scoring, never provides instructions
    return InterpretiveReviewResponse(
        ok=True,
        classification="ConditionallyPermitted",
        risk_framing={
            "account_risk": "high",
            "data_risk": "high",
            "legal_risk": "medium"
        },
        historical_context="This device class has been subject to independent security research. Unauthorized modification may result in data loss, account restriction, or legal exposure.",
        authority_paths=[
            {
                "type": "OEM",
                "description": "Contact device manufacturer for authorized recovery",
                "required_docs": ["Ownership proof"]
            }
        ],
        compliance_notes="This assessment documents analysis only. No modification or circumvention is performed or advised."
    )

# ============================================================================
# API ENDPOINTS
# ============================================================================

@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {"status": "ok", "service": "forgeworks-core", "version": "1.0.0"}

@app.get("/api/v1/ready")
async def readiness_check():
    """Readiness check - returns service status and capabilities"""
    return {
        "status": "ready",
        "service": "forgeworks-core",
        "version": "1.0.0",
        "capabilities": [
            "device_analysis",
            "ownership_verification",
            "legal_classification",
            "audit_logging",
            "authority_routing"
        ]
    }

@app.post("/api/v1/device/analyze", response_model=DeviceAnalyzeResponse)
async def analyze_device(request: DeviceAnalyzeRequest):
    """
    Analyze device (non-invasive, read-only)
    
    This endpoint performs diagnostic analysis only.
    It does NOT execute any modifications, exploits, or bypasses.
    """
    try:
        result = mock_device_analyze(request.device_metadata)
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/v1/ownership/verify", response_model=OwnershipVerifyResponse)
async def verify_ownership(request: OwnershipVerifyRequest):
    """
    Verify ownership claim
    
    This endpoint verifies ownership through attestation.
    It does NOT bypass any locks or security features.
    """
    try:
        result = mock_ownership_verify(request)
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/v1/legal/classify", response_model=LegalClassifyResponse)
async def classify_legal_status(request: LegalClassifyRequest):
    """
    Classify legal status based on device scenario
    
    This endpoint classifies legal status for routing purposes.
    It does NOT provide legal advice or execution instructions.
    """
    try:
        result = mock_legal_classify(request)
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/v1/compliance/summary", response_model=ComplianceSummaryResponse)
async def get_compliance_summary(request: ComplianceSummaryRequest):
    """
    Generate complete compliance summary
    
    This endpoint aggregates device analysis, ownership, legal classification,
    and routing into a single compliance report.
    """
    try:
        result = mock_compliance_summary(request.device_id)
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/v1/interpretive/review", response_model=InterpretiveReviewResponse)
async def interpretive_review(
    request: InterpretiveReviewRequest,
    x_ownership_confidence: Optional[float] = Header(None)
):
    """
    Interpretive Review (Custodian Vault - Analysis Only)
    
    This endpoint evaluates legal ambiguity and precedent.
    It does NOT execute actions or provide procedural guidance.
    
    Access Requirements:
    - Ownership confidence ≥ 85
    - Custodian role
    - All activity is logged
    """
    # Gate check
    if request.ownership_confidence < 0.85:
        raise HTTPException(
            status_code=403,
            detail="Ownership confidence must be ≥ 85% for interpretive review"
        )
    
    try:
        result = mock_interpretive_review(request)
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/api/v1/audit/events")
async def get_audit_events(
    limit: int = Query(50, ge=1, le=1000),
    level: Optional[str] = Query(None),
    action: Optional[str] = Query(None)
):
    """
    Get audit log events
    
    Returns immutable audit trail entries.
    """
    # Mock implementation - replace with real audit log query
    return {
        "ok": True,
        "events": [
            {
                "timestamp": "2025-01-10T00:00:00Z",
                "actor": "system",
                "action": "device_analysis",
                "result": "Allowed",
                "device_id": "dev_001"
            }
        ],
        "total": 1
    }

@app.get("/api/v1/audit/export")
async def export_audit_log(device_id: str):
    """
    Export audit log as PDF
    
    Returns a compliance-ready PDF with audit trail.
    """
    # Mock implementation - replace with real PDF generation
    return JSONResponse({
        "ok": True,
        "message": "PDF export not yet implemented",
        "device_id": device_id
    })

@app.get("/api/v1/route/authority")
async def route_to_authority(
    device_id: str,
    classification_status: str
):
    """
    Get authority routing guidance
    
    Returns OEM/carrier/court pathways based on classification.
    """
    route_map = {
        "Permitted": {"target": "OEM", "description": "Contact OEM service program"},
        "ConditionallyPermitted": {"target": "OEM", "description": "OEM authorization required"},
        "RequiresAuthorization": {"target": "LegalCounsel", "description": "Legal counsel required"},
        "Prohibited": {"target": "CourtSystem", "description": "Court order required"}
    }
    
    route = route_map.get(classification_status, {"target": "OEM", "description": "Standard routing"})
    
    return {
        "ok": True,
        "device_id": device_id,
        "route_to": route["target"],
        "description": route["description"],
        "required_documentation": ["Ownership proof", "Authorization documents"]
    }

@app.get("/api/v1/certification/status")
async def get_certification_status(user_id: Optional[str] = None):
    """
    Get certification status for current user
    
    Returns technician certification level and requirements.
    """
    return {
        "ok": True,
        "user_id": user_id or "current_user",
        "level": "Level I - Diagnostic Steward",
        "requirements_met": True,
        "next_level": "Level II - Repair Custodian"
    }

@app.get("/api/v1/ops/metrics")
async def get_ops_metrics():
    """
    Get operations dashboard metrics
    
    Returns active units, audit coverage, escalations.
    """
    return {
        "ok": True,
        "active_units": 0,
        "audit_coverage": "100%",
        "escalations": 0,
        "compliance_rate": "100%"
    }

# ============================================================================
# MAIN
# ============================================================================

if __name__ == "__main__":
    import uvicorn
    port = int(os.getenv("PORT", 8001))
    uvicorn.run(app, host="0.0.0.0", port=port)